Fixed bug in average precision
git-svn-id: svn://svn.code.sf.net/p/mulan/code/branches/mulan-eval-refactor@196 90ebec12-01d7-49c3-8feb-d42dc2bd4ece

+++ mulan/src/mulan/evaluation/RankingBasedMeasures.java
@@ -102,7 +102,7 @@ public class RankingBasedMeasures {
             else {
                 double sum=0;
                 for (int j : trueIndexes) {
-                    int rankedAbove=0;
+                    double rankedAbove=0;
                     for (int k : trueIndexes)
                         if (ranks[k] <= ranks[j])
                             rankedAbove++;
